+++ internal/dslx/quic.go
@@ -13,6 +13,7 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/ooni/probe-cli/v3/internal/logx"
 	"github.com/ooni/probe-cli/v3/internal/measurexlite"
 	"github.com/ooni/probe-cli/v3/internal/model"
 	"github.com/ooni/probe-cli/v3/internal/netxlite"
@@ -88,7 +89,7 @@ func (f *quicHandshakeFunc) Apply(
 	serverName := f.serverName(input)
 
 	// start the operation logger
-	ol := measurexlite.NewOperationLogger(
+	ol := logx.NewOperationLogger(
 		input.Logger,
 		"[#%d] QUICHandshake with %s SNI=%s",
 		trace.Index,
