+++ internal/experiment/tlsmiddlebox/dns.go
@@ -8,6 +8,7 @@ import (
 	"context"
 	"time"
 
+	"github.com/ooni/probe-cli/v3/internal/logx"
 	"github.com/ooni/probe-cli/v3/internal/measurexlite"
 	"github.com/ooni/probe-cli/v3/internal/model"
 )
@@ -17,7 +18,7 @@ func (m *Measurer) DNSLookup(ctx context.Context, index int64, zeroTime time.Tim
 	logger model.Logger, domain string, tk *TestKeys) ([]string, error) {
 	url := m.config.resolverURL()
 	trace := measurexlite.NewTrace(index, zeroTime)
-	ol := measurexlite.NewOperationLogger(logger, "DNSLookup #%d, %s, %s", index, url, domain)
+	ol := logx.NewOperationLogger(logger, "DNSLookup #%d, %s, %s", index, url, domain)
 	// TODO(DecFox, bassosimone): We are currently using the DoH resolver, we will
 	// switch to the TRR2 resolver once we have it in measurexlite
 	// Issue: https://github.com/ooni/probe/issues/2185
