+++ internal/logx/operation.go
@@ -1,8 +1,4 @@
-package measurexlite
-
-//
-// Logging support
-//
+package logx
 
 import (
 	"fmt"
@@ -11,8 +7,6 @@ import (
 
 	"github.com/ooni/probe-cli/v3/internal/model"
 )
-
-// TODO(bassosimone): consider moving inside the logx package?
 
 // NewOperationLogger creates a new logger that logs
 // about an in-progress operation. If it takes too much
