+++ internal/oohelperd/dns.go
@@ -9,7 +9,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/ooni/probe-cli/v3/internal/measurexlite"
+	"github.com/ooni/probe-cli/v3/internal/logx"
 	"github.com/ooni/probe-cli/v3/internal/model"
 	"github.com/ooni/probe-cli/v3/internal/netxlite"
 	"github.com/ooni/probe-cli/v3/internal/tracex"
@@ -55,7 +55,7 @@ func dnsDo(ctx context.Context, config *dnsConfig) {
 	defer reso.CloseIdleConnections()
 
 	// perform and log the actual DNS lookup
-	ol := measurexlite.NewOperationLogger(config.Logger, "DNSLookup %s", config.Domain)
+	ol := logx.NewOperationLogger(config.Logger, "DNSLookup %s", config.Domain)
 	addrs, err := reso.LookupHost(ctx, config.Domain)
 	ol.Stop(err)
